Chore: move exception util package
